requirements-dev.txt
-r requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0
black>=23.0.0
flake8>=6.1.0
